requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcc==0.18.0
black==23.11.0
blinker==1.9.0
Brlapi==0.8.3
cachetools==6.1.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.2
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
exceptiongroup==1.3.0
gitdb==4.0.12
GitPython==3.1.44
gyp==0.1
h11==0.16.0
hidpidaemon==18.4.6
httpcore==1.0.9
httplib2==0.20.2
httpx==0.25.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.3.0
isort==5.12.0
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kernelstub==3.1.4
keyring==23.5.0
langchain==1.2.6
langchain-core==1.2.7
langchain-openai==1.1.7
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libvirt-python==8.0.0
louis==3.20.0
macaroonbakery==1.3.1
MarkupSafe==3.0.2
more-itertools==8.10.0
mypy_extensions==1.1.0
narwhals==1.43.1
natsort==8.0.2
netaddr==0.8.0
netifaces==0.11.0
numpy==2.2.6
oauthlib==3.2.0
olefile==0.46
openai==2.15.0
orjson==3.11.5
ormsgpack==1.12.2
packaging==25.0
pandas==2.3.0
pathspec==1.0.3
pillow==11.2.1
platformdirs==4.5.1
pluggy==1.6.0
pop-transition==1.1.2
protobuf==6.31.1
pyarrow==20.0.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.12.5
pydantic-settings==2.1.0
pydantic_core==2.41.5
pydbus==0.6.0
pydeck==0.9.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==7.4.0
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==1.0.0
python-gnupg==0.4.8
python-telegram-bot==20.7
python-telegram-bot-calendar==1.0.5
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
referencing==0.36.2
regex==2026.1.15
repolib==2.2.1
repoman==1.4.0
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.25.1
SecretStorage==3.3.1
sessioninstaller==0.0.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.46.0
systemd-python==234
tenacity==9.1.2
tiktoken==0.12.0
toml==0.10.2
tomli==2.4.0
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==1.26.5
uuid_utils==0.14.0
wadllib==1.3.6
watchdog==6.0.0
wxPython==4.0.7
xdg==5
xkit==0.0.0
xxhash==3.6.0
zipp==1.0.0
zstandard==0.25.0
